move the native LLVM/LLD shim out of c/ and into std/llvm/native/
Six files, eight exported symbols, and the only C and C++ in the toolchain
that cannot become Salam: LLVM publishes InitializeAllTargets and friends
as macros, and LLD has no C API at all. They belong next to the std/llvm
bindings that declare them, not in the C compiler that is going away.
