+++ .github/workflows/compiler-selfhost.yml
@@ -275,8 +275,9 @@ jobs:
           path: libsalam_llvm.a
           key: >-
             salam-llvm-a-${{ matrix.os }}-llvm22-${{ hashFiles('c/Makefile',
-            'c/src/llvm/**', 'c/src/core/*.h', 'c/src/semantic/*.h',
-            'c/src/i18n/*.h', 'c/src/logger/*.h', 'c/src/diag/*.h') }}
+            'std/llvm/native/**', 'c/src/llvm/**', 'c/src/core/*.h',
+            'c/src/semantic/*.h', 'c/src/i18n/*.h', 'c/src/logger/*.h',
+            'c/src/diag/*.h') }}
 
       # Only useful on a miss, and only for the handful of objects that are
       # actually compiled (orc_call.c, lld_stub.c, lld_link.cc); the bulk of
