+++ .github/workflows/llvm-selfhost-gate.yml
@@ -25,6 +25,7 @@ on:
   pull_request:
     paths:
       - 'c/src/llvm/**'
+      - 'std/llvm/native/**'
       - 'c/src/semantic/**'
       - 'compiler/llvmgen/**'
       - 'compiler/semantic/**'
