+++ .gitignore
@@ -66,6 +66,9 @@ dist/
 # AI-facing documentation and the MCP server that serves it
 !docs/ai/AGENTS.md
 !tools/mcp/README.md
+# The native LLVM/LLD shim: why those six files are the one bit of C/C++
+# that stays after the compiler finished self-hosting.
+!std/llvm/native/README.md
 !tools/mcp/claude-plugin/skills/salam-lang/SKILL.md
 !tools/mcp/claude-plugin/commands/*.md
 # ---- Online Editor ----
